+++ src/lib.rs
@@ -22,10 +22,15 @@ pub mod instruction;
 use crate::instruction::Instruction;
 
 use display_interface::DataFormat;
-use display_interface::DisplayError;
 use display_interface::WriteOnlyDataCommand;
 use embedded_hal::blocking::delay::DelayUs;
 use embedded_hal::digital::v2::OutputPin;
+
+pub mod error;
+pub use error::Error;
+
+pub mod options;
+pub use options::*;
 
 pub mod models;
 use models::Model;
@@ -56,100 +61,6 @@ where
     orientation: Orientation,
     // Current MADCTL value
     madctl: u8,
-}
-
-///
-/// Display orientation.
-///
-#[derive(Debug, Clone, Copy)]
-pub enum Orientation {
-    /// Portrait orientation, with mirror image parameter
-    Portrait(bool),
-    /// Landscape orientation, with mirror image parameter
-    Landscape(bool),
-    /// Inverted Portrait orientation, with mirror image parameter
-    PortraitInverted(bool),
-    /// Inverted Lanscape orientation, with mirror image parameter
-    LandscapeInverted(bool),
-}
-
-impl Default for Orientation {
-    fn default() -> Self {
-        Self::Portrait(false)
-    }
-}
-
-impl Orientation {
-    pub fn value_u8(&self) -> u8 {
-        match self {
-            Orientation::Portrait(false) => 0b0000_0000,
-            Orientation::Portrait(true) => 0b0100_0000,
-            Orientation::PortraitInverted(false) => 0b1100_0000,
-            Orientation::PortraitInverted(true) => 0b1000_0000,
-            Orientation::Landscape(false) => 0b0010_0000,
-            Orientation::Landscape(true) => 0b0110_0000,
-            Orientation::LandscapeInverted(false) => 0b1110_0000,
-            Orientation::LandscapeInverted(true) => 0b1010_0000,
-        }
-    }
-}
-
-///
-/// Tearing effect output setting.
-///
-#[derive(Copy, Clone)]
-pub enum TearingEffect {
-    /// Disable output.
-    Off,
-    /// Output vertical blanking information.
-    Vertical,
-    /// Output horizontal and vertical blanking information.
-    HorizontalAndVertical,
-}
-
-///
-/// Defines expected color component ordering, RGB or BGR
-///
-#[derive(Debug, Clone, Copy)]
-pub enum ColorOrder {
-    Rgb,
-    Bgr,
-}
-
-impl Default for ColorOrder {
-    fn default() -> Self {
-        Self::Rgb
-    }
-}
-
-///
-/// Options for displays used on initialization
-///
-#[derive(Debug, Clone, Default)]
-pub struct DisplayOptions {
-    /// Initial display orientation (without inverts)
-    pub orientation: Orientation,
-    /// Set to make display vertical refresh bottom to top
-    pub invert_vertical_refresh: bool,
-    /// Specify display color ordering
-    pub color_order: ColorOrder,
-    /// Set to make display horizontal refresh right to left
-    pub invert_horizontal_refresh: bool,
-}
-
-///
-/// An error holding its source (pins or SPI)
-///
-#[derive(Debug)]
-pub enum Error<PE> {
-    DisplayError,
-    Pin(PE),
-}
-
-impl<PE> From<DisplayError> for Error<PE> {
-    fn from(_: DisplayError) -> Self {
-        Error::DisplayError
-    }
 }
 
 impl<DI, RST, M> Display<DI, RST, M>
